Add 'delete set' button
Server-side:
- add a DELETE handler for /rest/workout/exercise

Client-side:
- HTML templating for the [x] button, event handlers
- Also add "confirm?" dialog to the 'delete note' button

+++ static/app/app.js
@@ -154,6 +154,10 @@ define(['jquery', 'handlebars', 'bootstrap', 'd3', 'router', 'app/workout'],
             $("a#rm-note", this).each(function (idx) {
                 $(this).click(function (e) {
                     e.preventDefault();
+
+                    if (!confirm("OK to delete note?"))
+                        return;
+
                     $.ajax({ url: "/rest/note",
                              type: "DELETE",
                              data: { id: notes[idx].id },
